% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compas2d.R
\name{compas2d}
\alias{compas2d}
\title{Simulate a community matrix}
\usage{
compas2d(
  nsp = 40,
  le.grad = 100,
  ratio.grad = 0.8,
  nsite = 100,
  col_abund = NULL,
  row_abund = NULL,
  col_prefix = "s",
  row_prefix = "e",
  rowvar_prefix = "env",
  remove_zeroes = FALSE,
  buffer = 1,
  min.ab = 1,
  max.ab = 100,
  ninter = 100,
  delta = 1,
  mean.tol = 2,
  sd.tol = 10,
  return_intermediate = FALSE
)
}
\arguments{
\item{nsp}{Number of species / consumer}

\item{le.grad}{gradient length for traits/environmental variables. The values are
comprised between 0 and \code{le.grad}.}

\item{ratio.grad}{length of the second gradient as a ratio of the first one.}

\item{nsite}{Number of sites / resource species}

\item{col_abund}{Optional vector of abundances for species/consumers.
If missing, abundances are taken from a uniform law between \code{min.ab} and \code{max.ab}.}

\item{row_abund}{Optional vector of abundances in sites/ for resource species.
If missing, abundances are taken from a uniform law between \code{min.ab} and \code{max.ab}.}

\item{col_prefix}{Prefix for the column names of the matrix. Defaults to "s" for "species".}

\item{row_prefix}{Prefix for the row names of the matrix. Defaults to "e" for "environment".}

\item{rowvar_prefix}{Prefix for the column names of the row variable. Defaults to "env" for "environment".}

\item{remove_zeroes}{If there are unobserved species, should we keep them in the final matrix?}

\item{buffer}{buffer to allow species / consumer species optima to fall outside the gradient by
0 - \code{buffer} ; \code{le.grad} + \code{buffer}.}

\item{min.ab}{minimal abundance value for the uniform law (if \code{col_abund} and/or \code{row_abund} are absent)}

\item{max.ab}{maximal abundance value for the uniform law (if \code{col_abund} and/or \code{row_abund} are absent)}

\item{ninter}{Total number of observations in the web}

\item{delta}{Exponent between 0 and 1 to give more (1) or less (0) weight to trait matching relatively to
abundance.}

\item{mean.tol}{Mean value for the species/consumers niche tolerance (standard deviation).
The species/consumers niche tolerance are drawn from a normal law of mean \code{mean.tol}
and standard deviation \code{sd.tol} for each trait axis.}

\item{sd.tol}{Standard deviation for the species/consumers niche tolerance (standard deviation)
The species/consumers niche tolerance are drawn from a normal law of mean \code{mean.tol}
and standard deviation \code{sd.tol} for each trait axis.}

\item{return_intermediate}{Return intermediate results? If yes, will return the
probability of interaction based on matching only, on abundances only and the predicted
mix of matching and abundances (before sampling).}
}
\value{
An object of class compas, which is a list with 3 components (4 if \code{return_intermediate}):
\itemize{
\item comm: the community matrix of dimension \code{nsite} x \code{nsp}. Row names are \code{row_prefix}
followed by 1..\code{nsite} and column names are \code{col_prefix} followed by 1..\code{nsp}.
\item rowvar: the values of the row variables (resource traits/environment gradient values). It is a
nsite x 2 matrix. Row names are \code{row_prefix} followed by 1..\code{nsite} and column names are
\code{rowvar_prefix} followed by 1 and 2.
\item colvar: the values of the column variables (consumer/species traits values). It is a \code{nsp} x 2 x 2 array.
Row names are \code{row_prefix} followed by 1..\code{nsp}. Column names in each array dimension are \code{mean} and \code{sd} (for
the mean and standard deviation of the species niche, respectively). Each array 3rd dimension corresponds
to a resource trait or environmental gradient value.
\item if \code{return_intermediate}, a fourth component named intermediate is returned. The first element
is named \code{p_matching} and contains the matrix of probabilities of interactions based only on matching.
The second \code{ab_neutral} contains the matrix of count of interactions as predicted by species abundances. The third
\code{p_mix}  contains the matrix of probabilities of interactions taking into account the mix of matching and abundances.
}
}
\description{
Simulate abundance values based on species traits. The ecological interpretation can be
a sites x species matrix or a resource x consumer species interaction matrix.
}
