% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reciprocal.R
\name{correspondences}
\alias{correspondences}
\title{Get correspondences}
\usage{
correspondences(Pfreq0, Hrow, Hcol)
}
\arguments{
\item{Pfreq0}{Correspondence table}

\item{Hrow}{coordinate to use for rows (l x k)}

\item{Hcol}{coordinate to use for column (c x k)}
}
\value{
A table with k + 3 columns, where the k first columns are the correspondences coordinates
and columns \code{Row}, \code{Col} and \code{Weight} are added at the end.
}
\description{
Get the correspondences as a mean of the multivariate cordinates, diviede by sqrt(n)
.
}
