% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{create_code}
\alias{create_code}
\title{Create code}
\usage{
create_code(col, code_pattern = NA, name = NA)
}
\arguments{
\item{col}{column to code}

\item{code_pattern}{first letter to add before code}

\item{name}{optional name for output df}
}
\value{
dataframe with 2 columns: one with original vector, the other with the corresponding codes
}
\description{
Create a code for each (unique) value of a vector
}
