% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{df_to_matrix}
\alias{df_to_matrix}
\title{Dataframe to matrix}
\usage{
df_to_matrix(df, rows = "plant_species_code", columns = "animal_species_code")
}
\arguments{
\item{df}{the dataframe to transform. Must have column names indicated in \code{rows}
and \code{columns} and a third column \code{frequency} with the interaction frequency/strength index.}

\item{rows}{Name of the column with the species names that will be the rows of the matrix.}

\item{columns}{Name of the column with the species names that will be the columns of the matrix.}
}
\value{
Returns an incidence matrix with n rows and m columns filled with the values in \code{frequency}.
}
\description{
Transforms a dataframe into an incidence matrix.
}
