% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fd_niche.R
\name{fd.niche}
\alias{fd.niche}
\title{Functional diversity niche}
\usage{
fd.niche(
  networks,
  traits,
  splits = 30,
  dimensions = dim(pre.coord)[2],
  do.pcoa = TRUE,
  distance = "euclidean",
  calculate.centroid = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{networks}{network(s) that must be given as a list, even if there is only one network.
Each network in table format: consumer, resource, interaction strength
(if available, can be all 1 if not available).}

\item{traits}{resource species trait(s) that must be given as a list,
even if there is only one network.
Each data frame with species names as row names.}

\item{splits}{number of splits in each dimension for the grid used
to compute the convex hulls.}

\item{dimensions}{Number of dimensions to take into account}

\item{do.pcoa}{Do the PCoA?}

\item{distance}{Distance type to use}

\item{calculate.centroid}{Calculate centroids?}

\item{verbose}{Print information on the computations? (Added by Lisa Nicvert on 18/09/2024)}
}
\value{
A lsit copiling diversity measures.
Compile the results in two ways: per assemblage (FD_1) and per
index (FD_2).
For each element FD_1 and FD_2 several indices are provided.
Here I use only contrib.sum which gives the relative contribution of the species
to the sum of all species niches.
}
\description{
Function to calculate FD from traits and resource use (network data).
}
