% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{format_fit}
\alias{format_fit}
\title{Format fits}
\usage{
format_fit(matx, maty, idcols, bcodes = NULL)
}
\arguments{
\item{matx}{first matrix (values will be in values_x)}

\item{maty}{second matrix (values will be in values_x)}

\item{idcols}{columns to pivor}

\item{bcodes}{optional df of bird codes + names to merge names (codes must be named
"bird_species_code")}
}
\value{
A long dataframe merged version of the 2 matrices, where each row is an interaction and values
of the first and second matrix are respectively in "values_x" and "values_y"
}
\description{
Format fits matrices for plotting, i.e. pivot longer each
matrix and join them. Optionnally also add bird codes.
}
