% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{get_all_niche_measures}
\alias{get_all_niche_measures}
\title{Get niche breaths measures}
\usage{
get_all_niche_measures(
  mat,
  consumer_traits,
  resource_traits,
  nf_rec = 2,
  nf_PRN = nf_rec,
  splits_consumers = 100,
  splits_resources = 100,
  centroids = TRUE,
  ...
)
}
\arguments{
\item{mat}{interaction matrix}

\item{consumer_traits}{consumer traits matrix}

\item{resource_traits}{resource traits matrix}

\item{nf_rec}{axes to keep for reciprocal scaling}

\item{nf_PRN}{axes to keep for PRN}

\item{splits_consumers}{splits to do for resource traits}

\item{splits_resources}{splits for consumer traits}

\item{centroids}{return niche centroid measures too?}

\item{...}{additional arguments passed to \code{fd.niche}}
}
\value{
a list with 2 elements: \code{breadths} and \code{centroids} (or only \code{breadths}
if \code{centroids} is \code{FALSE}).
Each element contains a list of 2 dataframes: \code{rows} with row measures and \code{cols}
with column measures.
}
\description{
Returns niche breadths measures computed using different frameworks
}
