% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{get_best_model}
\alias{get_best_model}
\title{Compare 2 linear models}
\usage{
get_best_model(lmsimple, lm2, method = c("LRT", "AIC"), alpha = 0.05)
}
\arguments{
\item{lmsimple}{First model (if using method = "LRT", must be nested in lm2)}

\item{lm2}{Second model}

\item{method}{The method to use to compare models: likelihood ratio test
(use \code{method = "LRT"}) (for nested linear models comparison)
or AIC (use \code{method = "AIC"})}

\item{alpha}{Threshold to consider for the chi square test}
}
\value{
The best model
}
\description{
This function compares the models \code{lmsimple} and \code{lm2} and returns the best
}
