% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{get_cor_long}
\alias{get_cor_long}
\title{Get correlation table in its long form}
\usage{
get_cor_long(df)
}
\arguments{
\item{df}{The dataframe (all columns will be correlated).}
}
\value{
The dataframe with columns \code{colname}, \code{rowname} and \code{value}
(the correlation).
Some slots in \code{value} are \code{NA} and correspond to the repeated values.
\code{colname} and \code{rowname} are alphabetically ordered factors.
}
\description{
Get correlation table in its long form
}
