% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{get_mean_sd}
\alias{get_mean_sd}
\title{Get mean and variance}
\usage{
get_mean_sd(recscal, ax = 1:2)
}
\arguments{
\item{recscal}{A dataframe (expected to be the output of \code{reciprocal.coa} function).
The first columns must contain the reciprocal scaling scores and the last 3 columns
are \code{Row}, \code{Col} and \code{Weight}.}

\item{ax}{the axes for which to compute mean and variance (the function will use the n-th columns from
the \code{recscal} dataframe).}
}
\value{
A list of 4 matrices, each containing the mean or standard deviation per axes. Each matrix
has one column per axis given in \code{ax} and as many rows as there are grouping levels.
\itemize{
\item rowsd is the standard deviation per row
\item rowmean is the mean per row
\item colsd is the standard deviation per column
\item colmean is the mean per column
}

Rownames are the row or column groups. Column names are the same as in \code{recscal}.
}
\description{
Return the mean and variances from a reciprocal scaling analysis
}
