% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reciprocal.R
\name{get_mean_sd_RC}
\alias{get_mean_sd_RC}
\title{Get mean and standard deviation from RC/R/C scores}
\usage{
get_mean_sd_RC(RCscores, nax = 2, from = c("mix", "unique"))
}
\arguments{
\item{RCscores}{The output of reciprocal scaling. A list with components R, C and RC.}

\item{nax}{Number of axes to take into account}

\item{from}{Compute the scores from mixed scores (RC) or from the R scores (row scores, used
used for the mean/var of columns) and C scores (same in reverse)}
}
\value{
A list of 2 tables: dfli and dfco with the mean and variances per species for each
axis up to nax. The column names are "species" (species for which the mean/sd is computed),
meanxx (xx is the axis) and sdxx.
}
\description{
Compute a weighted mean or standard deviation from the R, C or RC scores
from reciprocal scaling.
}
