% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{get_PRN}
\alias{get_PRN}
\title{Get PRN niches}
\usage{
get_PRN(
  mat,
  resource_traits,
  consumer_traits,
  splits_consumers = 100,
  splits_resources = 100,
  dimensions,
  centroids = TRUE,
  ...
)
}
\arguments{
\item{mat}{interaction matrix (resources as rows x consumers as columns)}

\item{resource_traits}{resource traits matrix}

\item{consumer_traits}{consumer traits matrix}

\item{splits_consumers}{number of splits for resource niches}

\item{splits_resources}{number of splits for consumers niches}

\item{dimensions}{the number of dimensions for the convex hull}

\item{centroids}{return niche centroids}

\item{...}{Additional arguments passed to \code{fd.niche}}
}
\value{
a list with 2 elements: \code{breadths} and \code{centroids} (or only \code{breadths}
if \code{centroids} is \code{FALSE}).
Each element contains a list of 2 dataframes: \code{rows} with row measures and \code{cols}
with column measures.
}
\description{
Get PRN niches
}
