% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{get_recscal_niche}
\alias{get_recscal_niche}
\title{Get reciprocal scaling niche indices}
\usage{
get_recscal_niche(recscal, nfkeep = 3)
}
\arguments{
\item{recscal}{reciprocal scaling object}

\item{nfkeep}{number of axes to keep in the result.
This number determines the number of axes coordinates returned as well as
the number of axes used in the volumen computation.}
}
\value{
a list with 2 elements: \code{breadths} and \code{centroids} (or only \code{breadths}
if \code{centroids} is \code{FALSE}).
Each element contains a list of 2 dataframes: \code{rows} with row measures and \code{cols}
with column measures.
The elements of \code{breadths} have \code{nfkeep} + 2 columns: the niche breadths on each axis
(see \link{get_mean_sd} for details), the area on the first 2 axes and the volume on the first
\code{nfkeep} axes.
}
\description{
Get reciprocal scaling niche indices
}
