% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fd_niche.R
\name{inhull}
\alias{inhull}
\title{In hull}
\usage{
inhull(
  testpts,
  calpts,
  hull = geometry::convhulln(calpts),
  tol = mean(mean(abs(calpts))) * sqrt(.Machine$double.eps)
)
}
\arguments{
\item{testpts}{nxp array to test, n data points, in p dimensions.
If you have many points to test, it is most efficient to call this
function once with the entire set.}

\item{calpts}{mxp array of vertices of the convex hull, as used by convhulln.}

\item{hull}{(OPTIONAL) tessellation (or triangulation) generated by convhulln.
If hull is left empty or not supplied, then it will be generated.}

\item{tol}{tol - (OPTIONAL) tolerance on the tests for inclusion in the convex hull.
You can think of tol as the distance a point may possibly lie
outside the hull, and still be perceived as on the surface
of the hull.
Because of numerical slop nothing can ever be done exactly here.
I might guess a semi-intelligent value of tol to be
tol = 1.e-13*mean(abs(calpts(:)))
In higher dimensions, the numerical issues of floating point
arithmetic will probably suggest a larger value of tol.
In this R implementation default tol=mean(mean(abs(calpts)))*sqrt(.Machine$double.eps)
DEFAULT: tol = 1e-6}
}
\value{
Matlab returns a vector of TRUE (inside/on) or FALSE
(outside)
This R implementation returns an integer vector of length n
1 = inside hull
-1 = inside hull                                ANMERKUNG MATTHIAS: "outside"?
0 = on hull (to precision indicated by tol)
}
\description{
R implementation of the Matlab code by John D'Errico 04 Mar 2006
(Updated 30 Oct 2006) with some modifications and simplifications
}
\details{
Efficient test for points inside a convex hull in n dimensions

Keith Jewell
}
