% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{lm_labels}
\alias{lm_labels}
\title{Labels of a linear model}
\usage{
lm_labels(mod, a)
}
\arguments{
\item{mod}{The linear model. It is expected to have 2 or 3 coefficients of the form y = ax + b or y = ax + cx^2 + b.}

\item{a}{The axis to consider. Used for the subscript of the variables.}
}
\value{
A dataframe with columns \code{formula} and \code{r2} Containing respectively the model equation and coefficient of determination.
The formulas are written to be parsed later in the plot.
The variables' names are s_a^2 for y and m_a for x.
}
\description{
Get the labels of a linear model to display on a plot. Inspired from https://r-graphics.org/recipe-annotate-facet.
}
