% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{matrix_to_df}
\alias{matrix_to_df}
\title{Interaction matrix to dataframe}
\usage{
matrix_to_df(mat, colnames = "animals", rownames = "plants", tofactor = TRUE)
}
\arguments{
\item{mat}{Interaction matrix. The function assumes this matrix has row and
column names that are then used to fill the output dataframe columns.}

\item{colnames}{The name to give to the dataframe column containing the matrix column names.}

\item{rownames}{The name to give to the dataframe column containing the matrix row names.}

\item{tofactor}{Should the final columns \code{colnames} and \code{rownames} be of type factor?}
}
\value{
A dataframe with 3 columns:
\itemize{
\item \code{colnames} (defaults to "animals"), containing the initial matrix column names
\item \code{rownames} (defaults to "plants"), containing the initial matrix row names
\item \code{value} containing the matrix values
}
}
\description{
Transform interaction matrix (incidence matrix) to a dataframe with
3 columns: the names given in the arguments \code{colnames} and \code{rownames} and
a third column \code{value}.
}
