% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reciprocal.R
\name{nonrecscal}
\alias{nonrecscal}
\title{Non-reciprocal scaling}
\usage{
nonrecscal(dudi, oritab)
}
\arguments{
\item{dudi}{a \code{dudi} object expected to be of class \code{coa}, \code{pcaiv} or \code{dpcaiv}.}

\item{oritab}{original data table (matrix or data.frame). It must be a counts matrix
(it is converted to relative frequencies inside the function)}
}
\value{
A list with 2 components \code{H1} and \code{H2}, containing correspondences respectively
in the rows (scaling 1) and the columns space (scaling 2). For detains, see the correspondences
private function.
}
\description{
Preform non-reciprocal scaling and returns the correspondences coordinates in
rows and columns multivariate spaces.
}
