% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_figure_1.R
\name{patch_plot}
\alias{patch_plot}
\title{Patch plot}
\usage{
patch_plot(
  p_data,
  p_resource,
  p_consumer,
  n_consumer,
  n_resource,
  traits_dim = 1
)
}
\arguments{
\item{p_data}{data matrix plot (generated with \code{plot_data()})}

\item{p_resource}{resource traits plot (generated with \code{plot_traits()})}

\item{p_consumer}{consumer traits plot (generated with \code{plot_traits()})}

\item{n_consumer}{number of resources for plot height}

\item{n_resource}{number of consumers for plot width}

\item{traits_dim}{width/height of the traits matrix (use 1 for ordination, 2 for data)}
}
\value{
A patchwork ggplot
}
\description{
Arrange plots with patchwork
}
