% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_corcircle}
\alias{plot_corcircle}
\title{Plot correlation circle}
\usage{
plot_corcircle(
  cor,
  cor2 = NULL,
  xax = 1,
  yax = 2,
  xlab = NULL,
  ylab = NULL,
  label = rownames(cor),
  eig = NULL,
  mar = 0.01,
  labsize = 3.88,
  col_bg = "grey30",
  col = "black",
  lty = "solid",
  col2 = "black",
  lty2 = "solid",
  label2 = rownames(cor2),
  lim = 1,
  clip = "off"
)
}
\arguments{
\item{cor}{dataframe containing the coordinates of the variables for each axis in columns}

\item{cor2}{optional second set of variables (for double constrained analyses)}

\item{xax}{index of the column of the matrix to use for x}

\item{yax}{index of the column of the matrix to use for y}

\item{xlab}{Custom x-label}

\item{ylab}{Custom y-label}

\item{label}{label to display on the arrows}

\item{eig}{Eigenvalues vector}

\item{mar}{margin between arrow tips and label}

\item{labsize}{Size for arrow labels.}

\item{col_bg}{color to use for background elements (circle and y = 0 and x = 0)}

\item{col}{color for arrows (unique)}

\item{lty}{linetype for the first set of variables}

\item{col2}{optional color for the second set of variables}

\item{lty2}{linetype for the second set of variables}

\item{label2}{optional label to display on the arrows for the second set of variables}

\item{lim}{limits for x and y: a unique number that gives the distance detween the plot
limit and zero (so that the circle is centered)}

\item{clip}{constrain the labels to be inside the plot? If yes, change to "on".}
}
\value{
A ggplot with the correlation circle where variables are represented as arrows
inside the circle of radius 1.
}
\description{
Plot the correlation circle
}
