% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_lm_mean_sd}
\alias{plot_lm_mean_sd}
\title{Plot linear model}
\usage{
plot_lm_mean_sd(
  df,
  dat_pred,
  lab = NULL,
  col = "black",
  ylab = "Standard deviation",
  xlab = "Mean",
  text_size = 3,
  points_size = 1,
  facet = NULL,
  facet_dir = c("h", "v"),
  strip.position = "top",
  nudge_x = NULL,
  max.overlaps = 5
)
}
\arguments{
\item{df}{The dataframe containing the true data. It must have columns \code{mean} (used for the x-axis) and
\code{sd} (y-axis).}

\item{dat_pred}{The dataframe containing the predicted data (assumed to be over the true data range).
It is assumed to be obtained with \code{\link[=get_pred]{get_pred()}}. It must have columns \code{x}, \code{fit}, \code{lwr}, \code{upr}.}

\item{lab}{A label to add to the graph. If provided, will be added in the top-left corner.
It must be a dataframe with at least one column \code{R2} (to create this dataframe, see \link{lm_labels}).
If \code{facet} is not \code{NULL}, it must have a column
\code{type} too. If it has a column \code{formula}, the formula will be displayed along with the coefficient of determination.}

\item{col}{Color of the data points and prediction line. If there are facets, it should have length 2 (one color per
facet).}

\item{ylab}{Label of the y-axis}

\item{xlab}{Label of the x-axis}

\item{text_size}{Size of the text (points labels and R squared). Other text (axes titles, panels)
is scaled to be bigger.}

\item{points_size}{Size of the points}

\item{facet}{A column to use for facetting. Defaults to \code{NULL} (no facets).}

\item{facet_dir}{Direction of the facet grid: "h" for "horizontal and "v" for "vertical"
(useful only if \code{facet} is not \code{NULL})}

\item{strip.position}{Strip position (useful only if \code{facet} is not \code{NULL}). See the documentation
of \code{ggplot2::facet_wrap} (argument \code{strip.position}).}

\item{nudge_x}{Nudge to use for top left corner labels. If \code{NULL}, defaults to
1 \% of the range.}

\item{max.overlaps}{\code{max.overlaps} argument for \code{ggrepel::geom_text_repel}}
}
\value{
A ggplot object showing variance vs mean along with the predicted values of the linear model
}
\description{
Plot the linear models that predicts standard deviation with mean of the
reciprocal scaling.
}
