% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_matrix}
\alias{plot_matrix}
\title{Plot matrix}
\usage{
plot_matrix(
  mat,
  tile_fill = "white",
  col = "black",
  alpha = FALSE,
  max_size = max(mat) * 0.7,
  base_size = 12,
  xlab = "Animals",
  ylab = "Plants",
  trans = "identity",
  breaks = waiver(),
  legend_title = "Frequency"
)
}
\arguments{
\item{mat}{The matrix. It represents the interaction frequency
of species interactions (it is assumed to be plants x animals).}

\item{tile_fill}{Color fill of tiles. Use NA for no fill.}

\item{col}{Color of the points}

\item{alpha}{If true, will make less abundant interactions more
transparent}

\item{max_size}{The maximum circle area to be used in \code{scale_size_area}. Defaults to 0.7 times
the actual maximum.}

\item{base_size}{Base text size}

\item{xlab}{Title of the x axis}

\item{ylab}{Title of the y axis}

\item{trans}{Transformation to apply to the scaling of points.}

\item{breaks}{Breaks to use for the legend of the points size.}

\item{legend_title}{Title of the legend}
}
\value{
A ggplot
}
\description{
Plot an interaction matrix
}
