% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_niche}
\alias{plot_niche}
\title{Plot the niche}
\usage{
plot_niche(
  rec,
  sph,
  Y,
  mar = c(1, 1),
  step = 0.01,
  point_type = c("CA", "rec"),
  plot_all_points = FALSE,
  constant_point_size = FALSE,
  mary = 0.2,
  ntype = "li",
  lwd = 1,
  max_size = 3,
  alpha = 0.3,
  niche_col = "cornflowerblue",
  partner_col = "darkred",
  plot_curve = TRUE,
  empty = TRUE
)
}
\arguments{
\item{rec}{associated reciprocal scaling}

\item{sph}{Species to highlight}

\item{Y}{Interaction matrix}

\item{mar}{Left and right margin (vector)}

\item{step}{Step for x (traits step)}

\item{point_type}{Type of the points to display for the species interacting with the one
for which the niche is displayed: either "CA"
(means of reciprocal scaling coordinates) or "rec" (reciprocal scaling
correspondences).}

\item{plot_all_points}{Plot all points? If \code{TRUE}, will plot all points (species or correspondences).
If \code{FALSE}, will plot only the points for the highlignted species.}

\item{constant_point_size}{Keep point sizes constant? If \code{TRUE}, all points are of size
\code{max_size}, else, point sizes are weighted.}

\item{mary}{Margin on y min/max values}

\item{ntype}{Niche type (li or co)}

\item{lwd}{Linewidth}

\item{max_size}{Max points sizes}

\item{alpha}{Transparency of highlighted curves}

\item{niche_col}{Niche color}

\item{partner_col}{Partner color}

\item{plot_curve}{Plot the density curve? If \code{FALSE}, will only plot a vertical bar
to indicate niche optima.}

\item{empty}{Hide all legends and use theme_void?}
}
\value{
A ggplot with curves for niches
}
\description{
Plot the niche
}
