% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_niche_measures}
\alias{plot_niche_measures}
\title{Plot niche measures}
\usage{
plot_niche_measures(
  df,
  index,
  text_size = 2.5,
  col = "black",
  max.overlaps = 10
)
}
\arguments{
\item{df}{dataframe with niche measures indices as columns}

\item{index}{index to compare (in y)}

\item{text_size}{text size}

\item{col}{color of the points and regression line}

\item{max.overlaps}{maximum number of overlapping points}
}
\value{
a ggplot
}
\description{
Plot the relationships between niche measures
}
