% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_figure_1.R
\name{plot_table}
\alias{plot_table}
\title{Plot table}
\usage{
plot_table(
  x,
  y,
  abundance_data,
  abund_table,
  label_consumer = NULL,
  img_consumer = NULL,
  WA = FALSE
)
}
\arguments{
\item{x}{consumers abundance vector}

\item{y}{resources abundance vector}

\item{abundance_data}{abundance data table: expected to have columns}

\item{abund_table}{table corresponding to abundance_data (to wide format = interaction matrix)
Consumer_sp, Resource_sp, Abundance and coa_c and coa_r of WA is TRUE.}

\item{label_consumer}{label of consumer species (only useful if WA is TRUE)}

\item{img_consumer}{image
of consumer species (only useful if WA is TRUE)}

\item{WA}{display the weighted averaging for one species?}
}
\value{
a ggplot
}
\description{
Plot a matrix
}
