% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_figure_1.R
\name{plot_traits}
\alias{plot_traits}
\title{Plot traits}
\usage{
plot_traits(traits, order, label, img, fill = "grey", type = "consumer")
}
\arguments{
\item{traits}{traits dataframe (expected to have a \code{size} column)}

\item{order}{order to use for the species}

\item{label}{species labels}

\item{img}{species image}

\item{fill}{fill color (use a unique color to ignore the size)}

\item{type}{either "resource" (row) or "consumer" (column)}
}
\value{
a ggplot
}
\description{
Plot trait matrices
}
