% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_figure_1.R
\name{plot_traits_data}
\alias{plot_traits_data}
\title{Plot traits data}
\usage{
plot_traits_data(traits, label, img, fill, type = "consumer", reorder = FALSE)
}
\arguments{
\item{traits}{traits dataframe (expected to have a \code{size} column)}

\item{label}{species labels}

\item{img}{species image}

\item{fill}{fill color (use a unique color to ignore the size)}

\item{type}{either "resource" (row) or "consumer" (column)}

\item{reorder}{should species be reordered by label?}
}
\value{
a ggplot
}
\description{
Plot trait data matrices
}
