% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotmat}
\alias{plotmat}
\title{Plot matrices for the model}
\usage{
plotmat(
  mar = 2,
  marl = 1.2,
  E = FALSE,
  T_ = FALSE,
  r,
  c,
  l = NULL,
  k = NULL,
  name_mat = TRUE,
  name_dim = TRUE,
  Yname = "italic(Y)",
  Ename = "italic(E)",
  Tname = "italic(T)",
  rname = "italic(r)",
  cname = "italic(c)",
  horizr = TRUE,
  plot.margin.x = 1,
  plot.margin.y = 1
)
}
\arguments{
\item{mar}{horizontal margin between matrices}

\item{marl}{vertical margin between matrices}

\item{E}{Plot matrix E (rows variables)?}

\item{T_}{Plot matrix T (columns variables)?}

\item{r}{number of rows of matrix Y}

\item{c}{number of columns of matrix Y}

\item{l}{number of columns of matrix E}

\item{k}{number of columns of matrix T}

\item{name_mat}{display matrices names?}

\item{name_dim}{display dimension names?}

\item{Yname}{name of the matrix Y (center matrix)}

\item{Ename}{name of the matrix E (row variables matrix)}

\item{Tname}{name of the matrix T (column variables matrix)}

\item{rname}{name of the dimension r}

\item{cname}{name of the dimension c}

\item{horizr}{display r dimension horizontally? (else, will be vertical)}

\item{plot.margin.x}{horizontal margin around plot}

\item{plot.margin.y}{vertical margin around plot}
}
\value{
a ggplot of matrix Y (and optionally E and T),
with their dimensions written above and aligned in a square where
the fourth-corner is missing.
}
\description{
Function to plot matrices Y, E (environment/resource traits) and T (species traits/consumer traits)/
This function plots the matrices arranged in a square, with the user-given proportions.
}
\examples{
plotmat(r = 5, c = 2)
}
