% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reciprocal_ade4.R
\name{reciprocal.coa}
\alias{reciprocal.coa}
\title{Reciprocal scaling for CA}
\usage{
reciprocal.coa(x)
}
\arguments{
\item{x}{output of dudi.coa}
}
\value{
Results of reciprocal scaling.
Each row corresponds to a correspondence in the original table (a nonzero occurrence).
The first columns give coordinates in the different dimensions
(Scorexx). Scorexx gives the coordinates of the correspondence in the multivariate space as given
by canonical correlation analysis.
Row and Col give the row and column this correspondence belongs to.
Weight gives its weight (the count of the original cell divided by the sum of the table).
}
\description{
Performs reciprocal scaling for CA outputs as described by Thioulouse & Chessel (1992)
}
