% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reciprocal.R
\name{recscal}
\alias{recscal}
\title{Reciprocal scaling}
\usage{
recscal(Y, E = NULL, T_ = NULL)
}
\arguments{
\item{Y}{Count table}

\item{E}{Environment table (for rows)}

\item{T_}{Traits table (for columns)}
}
\value{
A list with R, C and RC results. Each have Scorexx columns, Col, Row and Weight.
Each row corresponds to a correspondence in the original table (a nonzero occurrence).
Scorexx gives the coordinates of the correspondence in the multivariate space as given
by canonical correlation analysis.
Row and Col give the row and column this correspondence belongs to.
Weight gives its weight (the count of the original cell divided by the sum of the table).
}
\description{
Perform reciprocal scaling from a count table and optionally
2 other "environment" and "traits" tables.
}
