% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reciprocal.R
\name{recscal_dudi}
\alias{recscal_dudi}
\title{Reciprocal scaling from multivariate coordinates}
\usage{
recscal_dudi(dudi, Y, transpose = FALSE)
}
\arguments{
\item{dudi}{the multivariate object (expects class 'coa', 'pcaiv' or 'dpcaiv'
from ade4)}

\item{Y}{Count table (converted to relative frequencies inside the function)}

\item{transpose}{Whether to invert R and C in the final table (intended for cases
where multivariate analysis was performed on a transposed table, like CCA
using animal traits for a environment x species table)}
}
\value{
A list with R, C and RC results. Each have Scorexx columns, Col, Row and Weight.
Each row corresponds to a correspondence in the original table (a nonzero occurrence).
Scorexx gives the coordinates of the correspondence in the multivariate space
(it is in essence a duplicated version of the multivariate scores).
Row and Col give the row and column this correspondence belongs to.
Weight gives its weight (the count of the original cell divided by the sum of the table).
}
\description{
Perform reciprocal scaling from a multivariate object and
the initial table
}
