% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAnetwork-package.R
\docType{package}
\name{CAnetwork-package}
\alias{CAnetwork}
\alias{CAnetwork-package}
\title{CAnetwork: Ecological Networks Analysis with Correspondence Analysis}
\description{
Analyze trait matching and interaction niches in ecological networks using correspondence analyses.
}
\author{
\strong{Maintainer}: Lisa Nicvert \email{lisa.nicvert@univ-lyon1.fr} (\href{https://orcid.org/0009-0006-5763-0865}{ORCID})

}
\keyword{internal}
