% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{multiplot}
\alias{multiplot}
\title{Plot bi, tri or quadriplots}
\usage{
multiplot(
  indiv_row = NULL,
  indiv_col = NULL,
  indiv_row_lab = rownames(indiv_row),
  indiv_col_lab = rownames(indiv_col),
  var_row = NULL,
  var_col = NULL,
  var_row_lab = rownames(var_row),
  var_col_lab = rownames(var_col),
  row_color = "black",
  col_color = "black",
  eig = NULL,
  x = 1,
  y = 2,
  xlim = NULL,
  ylim = NULL,
  grad = 4,
  mult = 1,
  title = NULL,
  max.overlaps = 20,
  labsize = 3,
  alphapoints = 0.8
)
}
\arguments{
\item{indiv_row}{Matrix of individuals coordinates for rows (type \code{dudi$li})}

\item{indiv_col}{Matrix of individuals coordinates for columns (type \code{dudi$co})}

\item{indiv_row_lab}{Labels for row individuals}

\item{indiv_col_lab}{Labels for columns individuals}

\item{var_row}{Matrix of variables coordinates for rows (type \code{dudi$corR})}

\item{var_col}{Matrix of variables coordinates for columns (type \code{dudi$corQ})}

\item{var_row_lab}{Labels for row variables}

\item{var_col_lab}{Labels for columns variables}

\item{row_color}{Color for the row individuals or variables}

\item{col_color}{Color for the columns individuals or variables}

\item{eig}{Eigenvalues vector}

\item{x}{Axis to use for the x-axis}

\item{y}{Axis to use for the y-axis}

\item{xlim}{x-axis limits. Defaults to data range if not specified}

\item{ylim}{y-axis limits. Defaults to data range if not specified}

\item{grad}{graduations for the major.grid}

\item{mult}{factor to multiply the vectors (arrows on the plot)}

\item{title}{plot title}

\item{max.overlaps}{\code{max.overlaps} argument for \code{ggrepel::geom_text_repel}}

\item{labsize}{text size for point labels}

\item{alphapoints}{Transparency value for points}
}
\value{
a ggplot
}
\description{
Bi, tri or quadriplots from a multivariate analysis
}
